style: _core 패키지 commons 로 통합
