+++ animory/src/main/java/com/daggle/animory/common/error/exception/Forbidden403.java
@@ -1,4 +1,5 @@
-package com.daggle.animory._core.errors.exception;
+package com.daggle.animory.common.error.exception;
+
 
 import com.daggle.animory.common.Response;
 import lombok.Getter;
@@ -7,17 +8,17 @@ import org.springframework.http.HttpStatus;
 
 // 유효성 검사 실패, 잘못된 파라메터 요청
 @Getter
-public class UnAuthorized401 extends RuntimeException {
+public class Forbidden403 extends RuntimeException {
 
-    public UnAuthorized401(String message) {
+    public Forbidden403(String message) {
         super(message);
     }
 
     public Response body(){
-        return Response.error(getMessage(), HttpStatus.UNAUTHORIZED);
+        return Response.error(getMessage(), HttpStatus.FORBIDDEN);
     }
 
     public HttpStatus status(){
-        return HttpStatus.UNAUTHORIZED;
+        return HttpStatus.FORBIDDEN;
     }
 }
