+++ animory/src/main/java/com/daggle/animory/common/error/exception/InternalServerError500.java
@@ -1,4 +1,4 @@
-package com.daggle.animory._core.errors.exception;
+package com.daggle.animory.common.error.exception;
 
 
 import com.daggle.animory.common.Response;
