+++ animory/src/main/java/com/daggle/animory/common/error/exception/UnAuthorized401.java
@@ -1,4 +1,4 @@
-package com.daggle.animory._core.errors.exception;
+package com.daggle.animory.common.error.exception;
 
 import com.daggle.animory.common.Response;
 import lombok.Getter;
@@ -7,17 +7,17 @@ import org.springframework.http.HttpStatus;
 
 // 유효성 검사 실패, 잘못된 파라메터 요청
 @Getter
-public class BadRequest400 extends RuntimeException {
+public class UnAuthorized401 extends RuntimeException {
 
-    public BadRequest400(String message) {
+    public UnAuthorized401(String message) {
         super(message);
     }
 
     public Response body(){
-        return Response.error(getMessage(), HttpStatus.BAD_REQUEST);
+        return Response.error(getMessage(), HttpStatus.UNAUTHORIZED);
     }
 
     public HttpStatus status(){
-        return HttpStatus.BAD_REQUEST;
+        return HttpStatus.UNAUTHORIZED;
     }
 }
